+++ docs/repod/changelog.rst
@@ -74,6 +74,7 @@ Fixed
   ``repod.file.package.Package.from_file`` were not done correctly, because
   after a previous ``MD5`` checksum calculation the package file was not read
   in its entirety.
+* Fix file mode validation for ``.MTREE`` files.
 
 Removed
 ^^^^^^^